requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
openai==1.3.7
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
litellm==1.44.0
